+++ pkg/services/ngalert/api/swagger-codegen/templates/partial_header.mustache
@@ -1,6 +1,6 @@
-/* Package api contains base API implementation of unified alerting
+/*Package api contains base API implementation of unified alerting
  *
- * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
+ *Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
  *
- * Do not manually edit these files, please find ngalert/api/swagger-codegen/ for commands on how to generate them.
+ *Do not manually edit these files, please find ngalert/api/swagger-codegen/ for commands on how to generate them.
  */
